+++ web/Examples/CustomControls/LiteMol-example.js
@@ -89,7 +89,8 @@ var LiteMol;
                         isoSigmaMax: 0,
                         radius: 5,
                         style: Visualization.Density.Style.create({
-                            isoSigma: -3,
+                            isoValue: -3,
+                            isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0xBB3333),
                             isWireframe: true,
                             transparency: { alpha: 1.0 }
@@ -102,7 +103,8 @@ var LiteMol;
                         isoSigmaMax: 5,
                         radius: 5,
                         style: Visualization.Density.Style.create({
-                            isoSigma: 3,
+                            isoValue: 3,
+                            isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0x33BB33),
                             isWireframe: true,
                             transparency: { alpha: 1.0 }
@@ -117,7 +119,8 @@ var LiteMol;
                         isoSigmaMax: 2,
                         radius: 5,
                         style: Visualization.Density.Style.create({
-                            isoSigma: 1.5,
+                            isoValue: 1.5,
+                            isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0x3362B2),
                             isWireframe: false,
                             transparency: { alpha: 0.45 }